+++ r/NEWS.md
@@ -10,7 +10,6 @@ Contributions welcome :)
 
 ## Breaking changes
 - Change first parameter name of `cluster_graph_leiden()`, `cluster_graph_louvain()` and `cluster_graph_seurat()` from `snn` to `mat` to more accurately reflect the input type.  (pull request #189)
-- Added parameters, `knn_method` and `knn_to_graph_method` to `cluster_graph_leiden()`, `cluster_graph_louvain()` and `cluster_graph_seurat()`. (pull request #189)
 
 ## Features
 - Add `write_matrix_anndata_hdf5_dense()` which allows writing matrices in AnnData's dense format, most commonly used for `obsm` or `varm` matrices. (Thanks to @ycli1995 for pull request #166)
@@ -19,7 +18,7 @@ Contributions welcome :)
 - Add feature selection functions `select_features_variance()`, and `select_features_{dispersion,mean,binned_dispersion}()`, with parameterization for normalization steps, and number of variable features (pull request #189)
 - Add `LSI()` and `IterativeLSI()` dimensionality functions to perform latent semantic indexing on a matrix (pull request #189).
 - Add capability to create partial function objects in when excluding the first argument of a function.  This is implemented in normalizations, feature selections, dimensionality reductions, and clustering functions. See `select_features_variance()` for usage.  (pull request #189)
-- Allowed clustering functions `cluster_graph_leiden()`, `cluster_graph_louvain()`, and `cluster_graph_seurat()` to also perform knn object and graph adjacency intermediate construction steps with `knn_method` and `knn_to_graph_method` parameters. Also provided `threads` and `verbose` arguments to clustering functions that are automatically passed down to knn object/graph adjacency construction steps.  Baseline expectation of usage in clustering functions is changed to having users directly input a cell embedding matrix. (pull request #189)
+- Create a wrapper function `cluster_cells_graph()` that wraps the steps of knn object creation, graph adjacency creation, and clustering all within a single function (pull request #189)
 
 ## Improvements
 - Speed up taking large subsets of large concatenated matrices, e.g. selecting 9M cells from a 10M cell matrix composed of ~100 concatenated pieces. (pull request #179)